requirement.txt
Django==2.0.8
djangorestframework==3.8.2
psycopg2==2.7.4
pytz==2019.2
sqlparse==0.3.0
